REFACTORING. Tried to add a structure for better continuous development. Not quite done yet.
