--- facebook/api.py
@@ -1,273 +1,0 @@
-#!/usr/bin/env python
-#
-# Copyright 2010 Facebook
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may
-# not use this file except in compliance with the License. You may obtain
-# a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations
-# under the License.
-from urllib2 import HTTPError
-
-"""Python client library for the Facebook Platform.
-
-This client library is designed to support the Graph API and the official
-Facebook JavaScript SDK, which is the canonical way to implement
-Facebook authentication. Read more about the Graph API at
-http://developers.facebook.com/docs/api. You can download the Facebook
-JavaScript SDK at http://github.com/facebook/connect-js/.
-
-If your application is using Google AppEngine's webapp framework, your
-usage of this module might look like this:
-
-    user = facebook.get_user_from_cookie(self.request.cookies, key, secret)
-    if user:
-        graph = facebook.GraphAPI(user["access_token"])
-        profile = graph.get_object("me")
-        friends = graph.get_connections("me", "friends")
-
-"""
-
-import cgi
-import hashlib
-import time
-import urllib
-import urllib2
-import httplib
-import base64
-import hmac
-
-import logging
-logger = logging.getLogger(__name__)
-
-# Find a JSON parser
-try:
-    import simplejson as json
-except ImportError:
-    try:
-        from django.utils import simplejson as json
-    except ImportError:
-        import json
-
-_parse_json = lambda s: json.loads(s)
-
-
-class GraphAPIError(Exception):
-    def __init__(self, type, message):
-        Exception.__init__(self)
-        self.type = type
-        self.message = message
-    
-    def __str__(self):
-        return '%s: %s' % (self.type, self.message)
-
-
-class GraphAPI(object):
-    """A client for the Facebook Graph API.
-
-    See http://developers.facebook.com/docs/api for complete documentation
-    for the API.
-
-    The Graph API is made up of the objects in Facebook (e.g., people, pages,
-    events, photos) and the connections between them (e.g., friends,
-    photo tags, and event RSVPs). This client provides access to those
-    primitive types in a generic way. For example, given an OAuth access
-    token, this will fetch the profile of the active user and the list
-    of the user's friends:
-
-       graph = facebook.GraphAPI(access_token)
-       user = graph.get_object("me")
-       friends = graph.get_connections(user["id"], "friends")
-
-    You can see a list of all of the objects and connections supported
-    by the API at http://developers.facebook.com/docs/reference/api/.
-
-    You can obtain an access token via OAuth or by using the Facebook
-    JavaScript SDK. See http://developers.facebook.com/docs/authentication/
-    for details.
-
-    If you are using the JavaScript SDK, you can use the
-    get_user_from_cookie() method below to get the OAuth access token
-    for the active user from the cookie saved by the SDK.
-    """
-    def __init__(self, access_token=None):
-        self.access_token = access_token
-
-    def get_object(self, id, **args):
-        """Fetchs the given object from the graph."""
-        return self.request(id, args)
-
-    def get_objects(self, ids, **args):
-        """Fetchs all of the given object from the graph.
-
-        We return a map from ID to object. If any of the IDs are invalid,
-        we raise an exception.
-        """
-        args["ids"] = ",".join(ids)
-        return self.request("", args)
-
-    def get_connections(self, id, connection_name, **args):
-        """Fetchs the connections for given object."""
-        return self.request(id + "/" + connection_name, args)
-
-    def put_object(self, parent_object, connection_name, **data):
-        """Writes the given object to the graph, connected to the given parent.
-
-        For example,
-
-            graph.put_object("me", "feed", message="Hello, world")
-
-        writes "Hello, world" to the active user's wall. Likewise, this
-        will comment on a the first post of the active user's feed:
-
-            feed = graph.get_connections("me", "feed")
-            post = feed["data"][0]
-            graph.put_object(post["id"], "comments", message="First!")
-
-        See http://developers.facebook.com/docs/api#publishing for all of
-        the supported writeable objects.
-
-        Most write operations require extended permissions. For example,
-        publishing wall posts requires the "publish_stream" permission. See
-        http://developers.facebook.com/docs/authentication/ for details about
-        extended permissions.
-        """
-        assert self.access_token, "Write operations require an access token"
-        return self.request(parent_object + "/" + connection_name, post_args=data)
-
-    def put_wall_post(self, message, attachment={}, profile_id="me"):
-        """Writes a wall post to the given profile's wall.
-
-        We default to writing to the authenticated user's wall if no
-        profile_id is specified.
-
-        attachment adds a structured attachment to the status message being
-        posted to the Wall. It should be a dictionary of the form:
-
-            {"name": "Link name"
-             "link": "http://www.example.com/",
-             "caption": "{*actor*} posted a new review",
-             "description": "This is a longer description of the attachment",
-             "picture": "http://www.example.com/thumbnail.jpg"}
-
-        """
-        return self.put_object(profile_id, "feed", message=message, **attachment)
-
-    def put_comment(self, object_id, message):
-        """Writes the given comment on the given post."""
-        return self.put_object(object_id, "comments", message=message)
-
-    def put_like(self, object_id):
-        """Likes the given post."""
-        return self.put_object(object_id, "likes")
-
-    def delete_object(self, id):
-        """Deletes the object with the given ID from the graph."""
-        return self.request(id, post_args={"method": "delete"})
-
-    def request(self, path, args=None, post_args=None):
-        """Fetches the given path in the Graph API.
-
-        We translate args to a valid query string. If post_args is given,
-        we send a POST request to the given path with the given arguments.
-        """
-        if not args: args = {}
-        if self.access_token:
-            if post_args is not None:
-                post_args["access_token"] = self.access_token
-            else:
-                args["access_token"] = self.access_token
-        if post_args:
-            for k, v in post_args.iteritems():
-                if isinstance(v, basestring):
-                    post_args[k] = v.encode('utf-8')
-        post_data = None if post_args is None else urllib.urlencode(post_args)
-        query = "https://graph.facebook.com/" + path + "?" + urllib.urlencode(args)
-        logger.debug('query: %s' % query)
-        try:
-            file = urllib2.urlopen(query, post_data)
-            raw = file.read()
-        except HTTPError as e:  # attrs: filename, code, msg, hdrs, fp
-            if e.fp is not None:
-                r = _parse_json(e.fp.read())
-                raise GraphAPIError(r['error']['type'], r['error']['message'])
-            else:
-                raise GraphAPIError('HTTP ERROR %s' % e.code, '%s, %s %s' % (e.msg, e.filename, e.fp.read()))
-        logger.debug('facebook response raw: %s, query: %s' % (raw, query))
-        try:
-            response = _parse_json(raw)
-            if isinstance(response, dict) and response.get("error"):
-                raise GraphAPIError(response["error"]["type"],
-                                    response["error"]["message"])
-        except AttributeError:
-            pass
-        finally:
-            file.close()
-            
-        return response
-
-
-def base64_url_decode(s):
-    return base64.urlsafe_b64decode(s.encode("utf-8") + '=' * (4 - len(s) % 4))
-
-
-def parseSignedRequest(signed_request, secret=None, application=None):
-    """
-    adapted from from
-    http://web-phpproxy.appspot.com/687474703A2F2F7061737469652E6F72672F31303536363332
-    https://github.com/facebook/python-sdk/commit/cb43c5a4a4b8c3e66264ed5508871b175f9c515f
-    """
-
-    if not secret:
-        app_dict = get_app_dict(application)
-        secret = app_dict['SECRET']
-    
-    try:
-        (encoded_sig, payload) = signed_request.split(".", 2)
-    except IndexError:
-        raise ValueError("Signed Request is malformed")
-    
-    sig = base64_url_decode(encoded_sig)
-    data = json.loads(base64_url_decode(payload))
-
-    if data.get("algorithm").upper() != "HMAC-SHA256":
-        raise ValueError("'signed_request' is using an unknown algorithm")
-    else:
-        expected_sig = hmac.new(secret, msg=payload, digestmod=hashlib.sha256).digest()
-
-    if sig != expected_sig:
-        raise ValueError("'signed_request' signature mismatch")
-    
-    return data
-
-
-# Currently not in use. And not working
-"""
-def get_user_from_cookie(cookies, app_id, app_secret):
-    ""Parses the cookie set by the official Facebook JavaScript SDK.
-
-    cookies should be a dictionary-like object mapping cookie names to
-    cookie values.
-
-    If the user is logged in via Facebook, we return a dictionary with the
-    keys "uid" and "access_token". The former is the user's Facebook ID,
-    and the latter can be used to make authenticated requests to the Graph API.
-    If the user is not logged in, we return None.
-
-    Download the official Facebook JavaScript SDK at
-    http://github.com/facebook/connect-js/. Read more about Facebook
-    authentication at http://developers.facebook.com/docs/authentication/.
-    ""
-    cookie = cookies.get("fbsr_" + app_id, "")
-    if not cookie: return None
-    parsed_request = parseSignedRequest(cookie, app_secret)
-    
-""" 
-    
-
